available workspace categories as selector
